# perf: file read performance -- double reads, extra copies