@lgritz 2024-10-28
Opinions?